formula.h is working

+++ S2/sample/include/stakk.h
@@ -2,29 +2,41 @@
 
 #include "arr.h"
 
-
-enum Operators {SUM, SUB, MUL, DIV, OPEN_BRACKET, CLOSE_BRACKET};
-
 template <typename T> class stakk {
 public:
+
+  stakk() {}
+
+  void print() {
+    for (int i = 0; i < m_storage.Size(); ++i)
+      cout << m_storage[i] << " ";
+    cout << endl;
+  }
 
   void push(const T& data) {
     m_storage.addData(m_storage, data);
   }
+
   size_t size()const {
     return m_storage.Size();
   }
+
   bool empty()const {
     return m_storage.isVoid();
   }
-  const T& looktop()const {
-    if (!m_storage.isVoid())
+
+  T looktop()const {
+    if (!empty())
       return  m_storage.getValue(m_storage.Size() - 1);
     else
       throw std::logic_error("Stack is empty!");
   }
+
   void pop() {
-    m_storage.deleteElement(m_storage.Size() - 1);
+    if (!empty())
+      m_storage.deleteElement(m_storage.Size() - 1);
+    else
+      throw std::logic_error("Stack is empty!");
   }
 
 private:
